refactor: move stack trace functionality into a separate module
